Word_World
This program takes input from the user for a text and a pattern. Then, it compares each word in the text with the pattern and prints the ones that match the pattern on the screen. The pattern can contain letters along with '*' (matches any number of letters) or '-' (matches only one letter) characters. Then, it uses regular expressions to check for matches between the pattern and the word. It prints the matching words on the screen.